+++ lib/models/new_expense.dart
@@ -1,0 +1,21 @@
+import 'package:kasa_w_grupie/models/expense.dart';
+
+class NewExpense {
+  String groupId;
+  String? name;
+  double? amount;
+  DateTime? date;
+  String? payer;
+  String? description;
+  ExpenseSplit? split;
+
+  NewExpense({
+    required this.groupId,
+    this.name,
+    this.amount,
+    this.date,
+    this.payer,
+    this.split,
+    this.description,
+  });
+}
