+++ lib/services/expense_service.dart
@@ -1,24 +1,4 @@
-import 'package:kasa_w_grupie/models/expense.dart';
-
-class NewExpense {
-  String groupId;
-  String? name;
-  double? amount;
-  DateTime? date;
-  String? payer;
-  String? description;
-  ExpenseSplit? split;
-
-  NewExpense({
-    required this.groupId,
-    this.name,
-    this.amount,
-    this.date,
-    this.payer,
-    this.split,
-    this.description,
-  });
-}
+import 'package:kasa_w_grupie/models/new_expense.dart';
 
 abstract class ExpenseService {
   Future<String?> addExpense(NewExpense expense);
